service/src/doc_agent/tools/file_module/requirements.txt
# 文件处理模块依赖包

# 核心依赖
requests>=2.25.1
chardet>=4.0.0

# 文档解析依赖
python-docx>=0.8.11
pandas>=1.3.0
openpyxl>=3.0.9
python-pptx>=0.6.21

# HTML解析依赖
beautifulsoup4>=4.9.3
lxml>=4.6.3

# 可选依赖（用于更好的性能）
xlrd>=2.0.1
